requirements.txt
flask
openai
gspread
python-dotenv
oauth2client